pyproject.toml
[project]
name = "passw-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "passw_manager", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
